+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Sat Nov 09 13:12:36 KST 2024
+gradle.version=8.10.2
