+++ src/main/java/org/seoultech/tableapi/menu/dto/MenuListResponse.java
@@ -1,6 +1,6 @@
-package org.seoultech.tableapi.domain.menu.dto;
+package org.seoultech.tableapi.menu.dto;
 
-import org.seoultech.tableapi.domain.menu.domain.Menu;
+import org.seoultech.tableapi.menu.domain.Menu;
 
 import java.util.List;
