+++ src/main/java/org/seoultech/tableapi/menu/repository/MenuRepository.java
@@ -1,6 +1,6 @@
-package org.seoultech.tableapi.domain.menu.repository;
+package org.seoultech.tableapi.menu.repository;
 
-import org.seoultech.tableapi.domain.menu.domain.Menu;
+import org.seoultech.tableapi.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuRepository extends JpaRepository<Menu, Long> {
