+++ src/main/java/org/seoultech/tableapi/menu/repository/MenuRepositoryCustom.java
@@ -1,6 +1,6 @@
-package org.seoultech.tableapi.domain.menu.repository;
+package org.seoultech.tableapi.menu.repository;
 
-import org.seoultech.tableapi.domain.menu.domain.Menu;
+import org.seoultech.tableapi.menu.domain.Menu;
 
 import java.util.List;
 import java.util.Optional;
