+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
+
 spring.application.name=table-api
